test audio!

+++ src/audio_out.h
@@ -7,13 +7,17 @@
 #ifndef AUDIO_OUT_H
 #define AUDIO_OUT_H
 
+#include "hardware/dma.h"
+#include "hardware/pwm.h"
+#include "pico/stdlib.h"
+
 class audio_out {
 public:
 	audio_out(uint pin, double sampling_freq);
 	~audio_out();
 
 	// play some audio asynchronously
-	void play(const uint8_t* samples, size_t length);
+	void play(const uint32_t* samples, size_t length);
 
 	// 
 	void wait();
